+++ TDM.Core/TDM/Server/Web/Program.cs
@@ -1,5 +1,7 @@
 #nullable enable
 
+using System;
+using System.Diagnostics;
 using System.Threading.Tasks;
 using Domain.Supermodel.Persistence;
 using Microsoft.AspNetCore.Hosting;
@@ -13,12 +15,16 @@ namespace Web
     {
         public static async Task Main(string[] args)
         {
-            //This removes and recreates the database
-            await using (new UnitOfWork<DataContext>())
+            if (Debugger.IsAttached || !await EFCoreUnitOfWorkContext.Database.CanConnectAsync())
             {
-                await EFCoreUnitOfWorkContext.Database.EnsureDeletedAsync();
-                await EFCoreUnitOfWorkContext.Database.EnsureCreatedAsync();
-                await UnitOfWorkContext.SeedDataAsync();
+                Console.Write("Recreating the database... ");
+                await using (new UnitOfWork<DataContext>())
+                {
+                    await EFCoreUnitOfWorkContext.Database.EnsureDeletedAsync();
+                    await EFCoreUnitOfWorkContext.Database.EnsureCreatedAsync();
+                    await UnitOfWorkContext.SeedDataAsync();
+                }
+                Console.WriteLine("Done!");
             }
 
             await CreateHostBuilder(args).Build().RunAsync();
